Stage 2/5: Rocking OAuth
Description

You have written a simulation of the final application, and now we will make it fully functional, step by step.

Working with OAuth
OAuth means a protocol for authorization. When you are visiting a website that allows you to register, you often have the option of signing in with Google, Twitter, Facebook, or another service. All these services use OAuth to provide user information to third-party applications.

The main entity of OAuth is access_token, the secret code that should be sent with an HTTP request to API, so that the service is sure that you have enough rights to get information from API. So, before an application can receive access_token, the user should confirm access to that application.
In our case, using the Spotify API, it will look like this:

https://accounts.spotify.com/authorize?client_id=YOURCLIENT&redirect_uri=https://www.example.com&response_type=code
1. The user follows this link and confirms access.
2. Then the user is redirected to redirect_uri with the response in the URL (http://www.example.com?code=7angkqw2DAsdfkQ). (Your own server waiting for code.)
3. You use this code to get access_token and make requests to an API.

Objectives
So, in this stage, you must go to the Spotify Web site for developers and create your application.
To create an application, you should select Dashboard tab on the site, log in to Spotify, and click the button Create an App.

Add one more command to your program: an auth command that will print the auth link and allow us to use another command (no real authorization required, just print a workable link and make some Boolean field true). Don't forget to put your client id in this link.

Make commands unavailable if user access for your program is not confirmed (if they did not call the auth command).

Remember that you should add your redirect_uri in the settings of your application.