Stage 5/5: Final chords
Description

Since the results can take up many lines, sometimes it is not convenient to see the whole result at once, and you might want to see everything separately. In this stage, you will need to reorganize your application and add paginated output.

MVC pattern
MVC (Model View Controller) is an architectural pattern for building an application. Using this pattern we achieve separation of data and presentation, scalability, and maintainability. The main concept of MVC is the division of the program into three parts. Model (or just data) is manipulated by the user performing various actions. Controller receives these actions from the user and interprets them, changes the model, and updates the view. The view determines how the elements on the screen will be displayed, what text they will have, indents, etc.

1. The user makes some actions that Controller intercepts.
2. The controller manipulates the Model according to the user’s request.
3. Then the Controller updates the View that the user sees.

Objectives
According to the MVC pattern, reorganize your application into three components (the controller should read input and make requests to the API and update the view) and create a paginated output that will display 5 entries per page. You should be able to navigate through the pages using the commands next (see the next page) and prev (see the previous page). If the user is viewing the last page or the first one and calls next or prev, then you need to display a message: "No more pages." and stay at the same page. Also, under each output should be a message containing the number of the current page and how many total pages there are.

Don't forget about -access and -resource arguments and add another argument: a number of entries that should be shown on a page. Your program should process argument -page. If ths argument isn't set, you should use the default value 5.